add a hardcode separate list of emotions
